Added 3D gameplay: * Dyzx now have a z component in every quantity * Because of that they .._x, .._y   variable pairs we replaced with genuine vectors (but to avoid overhead we do vector operations explicitly) * Added a perspective scale that enlarges the dyzk as it goes higher (i.e. gets closer to the camera) * In order to handle physics correctly, arena now feeds local normal/depth data to dyzx * Collision detection now takes into account z
From a physics point of view the perspective scale is a hack as things do not get bigger when they get closer
to the viewer, the viewer sees them that way. However in order to implement perspective view properly we need
to treat the arena image data we have differently (not to mention transforming the coordinates of every
physical object)and from rendering point of view it is much easier to pretend that the arena is
already "perspective" than treating it as "othogonal" and displacing every pixel to its
correct location. Long story short - this is a hack, but we are happy with it! HAPPY!!!